requirements.txt
discord.py==2.0.0
babel==2.10.3